fix setting of lines-changed-only
The README states that the value of true focuses on
lines in diff with additions only.
Setting it to 'diff' will use the original behavior (meaning
entire diff chunks with unchanged lines).

However, I forgot to change this value's
interpretation in the src code (& test).

This commit makes the README accurate.