fix: pin Python to 3.12
As of Ubuntu 24.04 LTS, the default version of Python is 3.12. Whereas, Ubuntu 22.04 LTS used Python 3.10, which is reaching End-of-Life next week.

Apparently, I already had created a .python-version file locally, but it was gitignored without my knowledge.

This should helps us control upgrades to Python version while some dependencies, namely `pygit2`, need extra time to add support for the new Python versions.

It also helps avoid disruption for users that are not aware of our use of a Python venv.